requirements.txt
ag2==0.9.9
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
applicationinsights==0.11.10
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.9
av==14.4.0
azure-ai-agents==1.2.0b3
azure-ai-projects==1.1.0b3
azure-cognitiveservices-speech==1.45.0
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==4.9.0
azure-identity==1.24.0
azure-search-documents==11.5.3
azure-storage-blob==12.26.0
backoff==2.2.1
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.1
cloudevents==1.12.0
colorama==0.4.6
croniter==6.0.0
cryptography==45.0.7
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
email-validator==2.3.0
faiss-cpu==1.12.0
fastapi==0.124.2
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.27.0
google-auth==2.41.1
google-crc32c==1.7.1
google-genai==1.42.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
ifaddr==0.2.0
importlib_metadata==8.7.0
iniconfig==2.3.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.1
langchain-google-genai==3.0.0
langchain-qdrant==1.1.0
langchain-text-splitters==1.0.0
langgraph==1.0.1
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.9
langsmith==0.4.38
lazy-object-proxy==1.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mem0ai==1.0.0
more-itertools==10.8.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.9.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.106.1
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
parse==1.20.2
pathable==0.4.4
pillow==11.3.0
pluggy==1.6.0
portalocker==3.2.0
posthog==6.8.0
prance==25.4.8.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybars4==0.9.13
pycparser==2.22
pydantic==2.12.5
pydantic-settings==2.10.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.1.0
pypdf==6.1.3
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
redis==5.0.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.1
rpds-py==0.27.1
rq==2.6.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scipy==1.16.1
semantic-kernel==1.36.2
sentry-sdk==2.42.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.50.0
streamlit==1.50.0
tenacity==9.1.2
termcolor==3.1.0
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
Werkzeug==3.1.1
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
